pagination backend, checkout handle, tax fixs

+++ app/Http/Controllers/ProdukController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Keranjang;
 use App\Models\Produk;
+use App\Models\Transaksi;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -49,6 +51,7 @@ class ProdukController extends Controller
     public function show(Produk $id) {
         return view('pembeli.show', [
             'product' => $id,
+            'keranjang' => Keranjang::all()->where('pembeli_id', Auth::user()->id)->count(),
         ]);
     }
     
@@ -56,6 +59,10 @@ class ProdukController extends Controller
         return view('penjual.update', [
             'product' => $produk,
         ]);
+    }
+
+    public function success() {
+        return view('pembeli.success');
     }
     
     public function update(Request $request, $id) {
@@ -81,14 +88,18 @@ class ProdukController extends Controller
         ]);
         $produk->update($validateData);
 
-        return redirect("/produk")->with('success', 'Rental berhasil diubah');
+        return redirect("/produk")->with('success', 'Produk berhasil diubah');
     }
     
     public function destroy($id) {
         $produk = Produk::findOrFail($id);
-        $produk->delete();
-
-        return redirect('/produk')->with('success', 'Rental berhasil dihapus');
+        $transaksi = Transaksi::where('produk_id', $produk->id)->first();
+        if($produk->stok == 0 || $transaksi) {
+            return redirect('/produk')->with('error', 'Produk tidak dapat dihapus');
+        } else {
+            $produk->delete();
+            return redirect('/produk')->with('success', 'Produk berhasil dihapus');
+        }
     }
 }
